@@ -2,31 +2,44 @@
   "name": "react-forms",
   "version": "0.1.0",
   "private": true,
-  "babel": {
-    "presets": [
-      "airbnb",
-      "env",
-      "stage-0",
-      "react"
-    ]
-  },
-  "devDependencies": {
-    "react-scripts": "^3.3.0",
-    "webpack": "^4.41.2",
-    "webpack-dev-server": "^3.9.0"
-  },
   "dependencies": {
-    "babel-core": "^6.26.3",
-    "babel-preset-airbnb": "^2.5.2",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-0": "^6.24.1",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.2"
+    "@testing-library/jest-dom": "^5.11.4",
+    "@testing-library/react": "^11.1.0",
+    "@testing-library/user-event": "^12.1.10",
+    "react": "^17.0.1",
+    "react-dom": "^17.0.1",
+    "react-scripts": "4.0.1",
+    "web-vitals": "^0.2.4"
   },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "dev:hot": "webpack-dev-server --hot --inline --progress --colors --watch --display-error-details --display-cached --content-base ./",
+    "test": "echo \".results.json\" && react-scripts test --reporters=@ihollander/jest-learn-reporter --reporters=default --watchAll",
     "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@ihollander/jest-learn-reporter": "^1.0.1",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.3.1",
+    "chai": "^4.2.0",
+    "enzyme": "^3.11.0",
+    "mocha": "^8.2.1"
   }
 }
