@@ -1,31 +1,15 @@
-import React from 'react';
+import React, { useState } from 'react';
 
-class Form extends React.Component {
-  state = {
-    firstName: "John",
-    lastName: "Henry"
-  }
+function Form() {
+  const [firstName, setFirstName] = useState("John")
+  const [lastName, setLastName] = useState("Henry")
 
-  handleFirstNameChange = event => {
-    this.setState({
-      firstName: event.target.value
-    })
-  }
-
-  handleLastNameChange = event => {
-    this.setState({
-      lastName: event.target.value
-    })
-  }
-
-  render() {
-    return (
-      <form>
-        <input type="text" name="firstName" onChange={event => this.handleFirstNameChange(event)} value={this.state.firstName} />
-        <input type="text" name="lastName" onChange={event => this.handleLastNameChange(event)} value={this.state.lastName} />
-      </form>
-    )
-  }
+  return (
+    <form>
+      <input type="text" value={firstName} />
+      <input type="text" value={lastName} />
+    </form>
+  )
 }
 
 export default Form;
